requirements.txt
# local package
-e .
#Python Version 3.9.8
# external requirements
protobuf==3.20.2 #Version that is compatibile with tensorflow and Feast
numpy==1.19.5 #Version that is compatibile with tensorflow and Feast
pandas==1.3.5 #Version that is compatibile with tensorflow and Feast
pandavro==1.5.2
tensorflow==2.6.2 #Tensorflow 2.6.2 require numpy=1.19 which is incompatible with FEAST 0.23.0 so had to lower the FEAST version that is compatible to numpy and protobuf 
matplotlib==3.5.3
sklearn==0.0.post1
category_encoders==2.5.1.post0
webcolors==1.11.1
lightgbm==3.2.1
seaborn==0.12.0
PyYAML==6.0
dill==0.3.6
cv==1.0.0
pyarrow==8.0.0
fastparquet==0.8.1
feast==0.19.3 #Release on 09-03-2022. Version is compatible with tensorflow, protobuf and numpy
typing-extensions==4.4.0
streamlit==1.16.0 

#Library used in the notebooks if required can be installed seperately
#cudf
#plotly

#Step 1 Create a Virtual enviroment with in the project folder
#py -<<python_version>> -m venv <<your_environment_name>>

#Step 2 Activate the Virtual Environment by calling Activate.bat 
#\<<your_environment_name>>\Scripts\Activate.bat

#Step 3: Select this Environment as Interpreatr in VScode
#-> Ctrl+Shift+P
#-> Select from drop down or type : "Python: Select Interpreter"
#-> Select "Enter interpreter path..."
#-> Select "Find.." and browse to folder and select" \Scripts\python.exe" in the new environment folder that we created.

#Step 4: [Optional]: Upgrade pip in your_enviroment
#-> Open the Terminal
#-> Terminal should show <<your_environment_name>> in the command line. If not execute Step 2 again
#-> pip install pip --upgrade

#Step 5: Install the requirement dll
#pip install -r requirements.txt
